+++ main.py
@@ -17,18 +17,7 @@ import sys
 import time
 
 from banner import banner
-
-
-class colors:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
-
+from colors import colors
 
 class InstagramOSINT:
